refactor: explicitly ensure appropriate signedness in settings
Explicitly ensure appropriate signedness when comparing integers for
width and height in src/settings.c for extreme edge cases.

+++ src/settings.c
@@ -482,8 +482,8 @@ int parse_size(int, const char **argv, void **references, int)
     if ((bitmask & HeightValue) == 0)
         height = width;
 
-    size->x = (width > INT_MAX) ? INT_MAX : width;
-    size->y = (height > INT_MAX) ? INT_MAX : height;
+    size->x = (width > (unsigned int) INT_MAX) ? INT_MAX : (int) width;
+    size->y = (height > (unsigned int) INT_MAX) ? INT_MAX : (int) height;
 
     return SUCCESS;
 }
